==> analyses/chromothripsis/01-process-sv-file.R <==
# Yang Yang 2019
# This script is for generating sv file in shatterseek-read/signature-read input format
# 
# input files are:
# 1.independent-specimens.wgs.primary-plus.tsv
# this file is used to choose independent specimens
# 2. pbta-sv-manta.tsv
# this file is the original sv file generated by manta call
# 
# output files are:
# sv files for each specimens
# 
# output file format:
# chrom1, pos1, chrom2, pos2, SVtype, strand1, strand2, alt1, alt2, homolen, inserlen
# chrom and pos are breakpoint location
# SVtype is structral variation type, they are DUP, DEL, INV and TRA 
# strand1 and strand2 will be used in shatterseek analysis
# alt1 and alt2 were used to infer strand， you can ignore them
# homolen and inserlen will be used in signature analysis

library(stringr) 
library(rprojroot)
library(readr)
library(dplyr)

# Detect the ".git" folder -- this will in the project root directory.
# Use this as the root directory to ensure proper sourcing of functions no
# matter where this is called from
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))

## =====================  Create A Subdirectory to Hold All The Output Files ===================== 
output_directory <- file.path(root_dir, "scratch","sv-vcf")
if (!dir.exists(output_directory)) {
  dir.create(output_directory, recursive = TRUE)
}

## ===================== Load  Independent Specimen List =====================
independent_specimen_list <-  read.table(file.path(root_dir, "data","independent-specimens.wgs.primary-plus.tsv"), header = TRUE, sep = "\t")
# bioid including all sample's names will be used later
bioid <- unique(independent_specimen_list$Kids_First_Biospecimen_ID)


## ===================== Load SV File =====================
# read sv file
# filter independent-specimens
# filter "PASS" samples
sv_analysis <- read_tsv(file.path(root_dir, "data","pbta-sv-manta.tsv.gz")) %>%
  filter((Kids.First.Biospecimen.ID.Tumor %in% bioid) & (FILTER == "PASS"))


## ===================== Functions =====================
extract_len <- function(info_field, len_string) {
  # info_field: an entry of the INFO column
  # len_string: ID of the INFO key that will be used to extract the length from the correct field
  len_string <-
    toupper(len_string)  #  forces len_string to uppercase letters
  if (str_detect(as.character(info_field), paste0(len_string, "="))) {
    return(gsub((";.*"), "", gsub((paste0(".*", len_string, "=")), "", as.character(info_field)
    )))
  } else {
    return(0)
  }
}

## ===================== Generate SV File In A Shatterseek-read Format =====================
for (i in bioid) {
  sv_individual <- sv_analysis[sv_analysis$Kids.First.Biospecimen.ID.Tumor  == i,]
  ## TYPE == BND
  # bnds is individual's translocation
  bnds <- sv_individual[sv_individual$SV.type == "BND", ]
  # delete the last two characters to match tranlocation pair
  reads_id_bnd <- substr(bnds$ID, start = 1, stop = nchar(as.character(bnds$ID))-2)
  # build bnds_file
  bnds_file <- matrix(NA, nrow = length(unique(reads_id_bnd)), ncol = 11)
  colnames(bnds_file) <- c("chrom1","pos1","chrom2","pos2","SVtype","strand1","strand2","alt1","alt2","homolen","inserlen")
  rownames(bnds_file) <- unique(reads_id_bnd)
   # if individual's translocation is null, skip next loop
  if (nrow(bnds)  != 0) {
    # fill bnds_file row by row
    for (j in 1:nrow(bnds)) {
      # extract reads id 
      # without the last two characters
      id_name <- substr(bnds[j, "ID"], start = 1, stop = nchar(as.character(bnds[j, "ID"])) - 2)
      # extract the last two chararcter of reads id
      number  <-  as.numeric(str_sub(bnds[j, "ID"], -1)) + 1
      bnds_file[id_name, paste0("chrom", number)] <- as.character(bnds$SV.chrom[j])
      bnds_file[id_name, paste0("pos", number)] <- as.character(bnds$SV.start[j])
      bnds_file[id_name, paste0("alt", number)] <- as.character(bnds$ALT[j])
      bnds_file[id_name, "homolen"] <- extract_len(bnds$INFO[j], "HOMLEN")
      bnds_file[id_name, "inserlen"] <- extract_len(bnds$INFO[j], "SVINSLEN")
      # estimate strands is + or -  by alt column
      # if alt column start with A/T/C/G， the strand is +, or else is -
      if (substr(bnds_file[id_name, paste0("alt", number)], start = 1, stop = 1) %in% c("A", "T", "C", "G")) {
        bnds_file[id_name, paste0("strand", number)] = "+"
      } else {
        bnds_file[id_name, paste0("strand", number)] = "-"
      }
    }
    bnds_file[, "SVtype"] <- "TRA"
  }
  ## TYPE == DUP
  # dups is individual's duplication
  dups <- sv_individual[sv_individual$SV.type == "DUP", ]
  reads_id_dup <- substr(dups$ID, start = 1, stop = nchar(as.character(dups$ID)))
  # build dups_file
  dups_file <- matrix(NA, nrow = length(unique(reads_id_dup)), ncol = 11)
  colnames(dups_file) <- c("chrom1","pos1","chrom2","pos2","SVtype","strand1","strand2","alt1","alt2","homolen","inserlen")
  rownames(dups_file) <- unique(reads_id_dup)
   # if individual's duplication is null, skip next loop
  if (nrow(dups)  != 0) {
    # fill dups_file row by row
    for (j in 1:nrow(dups)) { 
      id_name <- substr(dups[j, "ID"], start = 1, stop = nchar(as.character(dups[j, "ID"])))
      dups_file[id_name, "chrom1"] <- as.character(dups$SV.chrom[j])
      dups_file[id_name, "chrom2"] <- as.character(dups$SV.chrom[j])
      dups_file[id_name, "pos1"] <- as.character(dups$SV.start[j])
      dups_file[id_name, "pos2"] <- as.character(dups$SV.end[j])
      dups_file[id_name, "alt1"] <- as.character(dups$AnnotSV.ID[j])
      dups_file[id_name, "homolen"] <- extract_len(dups$INFO[j], "HOMLEN")
      dups_file[id_name, "inserlen"] <- extract_len(dups$INFO[j], "SVINSLEN")
      dups_file[id_name, "strand1"] = "-"
      dups_file[id_name, "strand2"] = "+"
    }
    dups_file[, "SVtype"] <- "DUP"
  }
  ## TYPE == DEL
  # dels is individual's deletion
  dels <- sv_individual[sv_individual$SV.type == "DEL", ]
  # extract all ids
  reads_id_del <- substr(dels$ID, start = 1, stop = nchar(as.character(dels$ID)))
  # build dels_file
  dels_file <- matrix(NA, nrow = length(unique(reads_id_del)), ncol = 11)
  colnames(dels_file) <- c("chrom1","pos1","chrom2","pos2","SVtype","strand1","strand2","alt1","alt2","homolen","inserlen")
  rownames(dels_file) <- unique(reads_id_del)
  # if individual's deletion is null, skip next loop
  if (nrow(dels)  != 0) {
    # fill dels_file row by row
    for (j in 1:nrow(dels)) { 
      id_name <- substr(dels[j, "ID"], start = 1, stop = nchar(as.character(dels[j, "ID"])))
      dels_file[id_name, "chrom1"] <- as.character(dels$SV.chrom[j])
      dels_file[id_name, "chrom2"] <- as.character(dels$SV.chrom[j])
      dels_file[id_name, "pos1"] <- as.character(dels$SV.start[j])
      dels_file[id_name, "pos2"] <- as.character(dels$SV.end[j])
      dels_file[id_name, "alt1"] <- as.character(dels$AnnotSV.ID[j])
      dels_file[id_name, "homolen"] <- extract_len(dels$INFO[j], "HOMLEN")
      dels_file[id_name, "inserlen"] <- extract_len(dels$INFO[j], "SVINSLEN")
      dels_file[id_name, "strand1"] = "+"
      dels_file[id_name, "strand2"] = "-"
    }
    dels_file[, "SVtype"] <- "DEL"
  }
  ## TYPE == INV
  # invs is individual's inversion
  invs <- sv_individual[sv_individual$SV.type == "INV", ]
  reads_id_inv <- substr(invs$ID, start = 1, stop = nchar(as.character(invs$ID)))
  # build invs_file
  invs_file <- matrix(NA, nrow = length(unique(reads_id_inv)), ncol = 11)
  colnames(invs_file) <- c("chrom1","pos1","chrom2","pos2","SVtype","strand1","strand2","alt1","alt2","homolen","inserlen")
  rownames(invs_file) <- unique(reads_id_inv)
   # if individual's inversion is null, skip next loop
  if (nrow(invs)  != 0) {
    # fill invs_file row by row
    for (j in 1:nrow(invs)) { 
      id_name <- substr(invs[j, "ID"], start = 1, stop = nchar(as.character(invs[j, "ID"])))
      invs_file[id_name, "chrom1"] <- as.character(invs$SV.chrom[j])
      invs_file[id_name, "chrom2"] <- as.character(invs$SV.chrom[j])
      invs_file[id_name, "pos1"] <- as.character(invs$SV.start[j])
      invs_file[id_name, "pos2"] <- as.character(invs$SV.end[j])
      invs_file[id_name, "alt1"] <- as.character(invs$AnnotSV.ID[j])
      invs_file[id_name, "homolen"] <- extract_len(invs$INFO[j], "HOMLEN")
      invs_file[id_name, "inserlen"] <- extract_len(invs$INFO[j], "SVINSLEN")
      if (str_detect(invs$INFO[j],"INV5")) {
        invs_file[id_name, "strand1"] = "+"
        invs_file[id_name, "strand2"] = "+"
        invs_file[id_name, "SVtype"] <- "h2hINV"
      } else {
        invs_file[id_name, "strand1"] = "-"
        invs_file[id_name, "strand2"] = "-"
        invs_file[id_name, "SVtype"] <- "t2tINV"
      }
    }
  }
  
  ## Merge four types SV
  sv_merge <- rbind(dups_file, dels_file, invs_file, bnds_file)
  sv_merge <- sv_merge[is.na(sv_merge[, 1]) == FALSE & is.na(sv_merge[, 3]) == FALSE, ]
  outputname_sv_merge <- paste0(i,".tsv")
  outputname_sv_merge  <- file.path(output_directory,outputname_sv_merge)
  
  ## Shatterseek only accept file with chrom1-22,X, so remove chrY and ChrM
  sv_merge_withoutY_withoutM <- sv_merge[sv_merge[,1] != "Y" & sv_merge[,3] != "Y" & sv_merge[,1] != "M" & sv_merge[,3] != "M",]
  outputname_sv_merge_withoutY_withoutM <- paste0(i,"_withoutYandM.tsv")
  outputname_sv_merge_withoutY_withoutM <- file.path(output_directory,outputname_sv_merge_withoutY_withoutM)
  
  ## output files
  write.table(sv_merge,file = outputname_sv_merge,sep = "\t",quote = FALSE,row.names = TRUE,col.names = NA)
  write.table(sv_merge_withoutY_withoutM,file = outputname_sv_merge_withoutY_withoutM,sep = "\t",quote = FALSE,row.names = TRUE,col.names = NA)
}
