==> analyses/chromothripsis/04-plot-chromothripsis-and-breakpoint-data.Rmd <==
---
title: 'Chromothripsis Plots with Breakpoint Data'
author: "Laura Egolf"
date: "2021"
output: html_document
---
  
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This notebook plots the number of chromothripsis regions per sample along with breakpoint data from `analyses/chromosomal-instability/`. As expected, the number of chromothripsis calls roughly correlates with the breakpoint counts calculated from both SV and CNV data. 

### Define directory paths and load packages

```{r}
# Set directory paths
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
analysis_dir <- file.path(root_dir, "analyses", "chromothripsis")
plots_dir <- file.path(analysis_dir, "plots", "04-breakpoint-data")
breakpoint_data_dir <- file.path(root_dir, "analyses", "chromosomal-instability", "breakpoint-data")

# Create plots_dir if it doesn't exist
if (!dir.exists(plots_dir)) {
  dir.create(plots_dir, recursive = TRUE)
}

# Load packages
library(ggplot2)
library(RColorBrewer)
library(readr)
library(dplyr)
library(grid)
```


### Prepare chromothripsis data and breakpoint data

```{r}
# Read in summary of chromothripsis regions per sample (number of regions for each confidence level)
chromoth_per_sample <- read_tsv(file.path(analysis_dir, "results", "chromothripsis_summary_per_sample.txt"))

# Read in CNV and SV breakpoint data
cnv_densities <- read_tsv(file.path(breakpoint_data_dir, "cnv_breaks_densities.tsv"))
sv_densities <- read_tsv(file.path(breakpoint_data_dir, "sv_breaks_densities.tsv"))

# Rename columns before merging
cnv_densities <- cnv_densities %>%
  rename(Kids_First_Biospecimen_ID = samples,
         cnv_breaks_count = breaks_count)
sv_densities <- sv_densities %>%
  rename(Kids_First_Biospecimen_ID = samples) %>%
  rename(sv_breaks_count = breaks_count)

# Merge chromothripsis data and breakpoint data
merge <- chromoth_per_sample %>% 
  inner_join(cnv_densities, by = "Kids_First_Biospecimen_ID") %>%
  inner_join(sv_densities, by = "Kids_First_Biospecimen_ID") 
```


### Plot the number of chromothripsis regions per sample along with CNV and/or SV break counts

Define plot theme:
```{r}
theme_set(theme_light())
```

Define color scale for # chromothripsis regions:
```{r}
max_regions <- max(merge$count_regions_any_conf)
colors <- colorRampPalette(brewer.pal(9,"YlOrRd"))(max_regions+1)
colors[1] <- "#1C00ff00"  # Assign no color to "0"
```

Truncate # chromothripsis regions above 5 (for stripcharts):
```{r}
merge <- merge %>%
  mutate(count_regions_any_conf_truncated = ifelse(count_regions_any_conf>=5, ">=5", count_regions_any_conf)) %>%
  mutate(count_regions_any_conf_truncated = as.factor(count_regions_any_conf_truncated)) %>%
  mutate(count_regions_any_conf_truncated = 
           forcats::fct_relevel(count_regions_any_conf_truncated, ">=5", after = Inf))
```


#### Scatterplot: CNV breaks vs. SV breaks, colored by # chromothripsis regions

```{r}
# Fit linear regression and create plot label
fit <- lm(sv_breaks_count ~ cnv_breaks_count, data = merge)
corr <- sqrt(summary(fit)$adj.r.squared)
pval <- summary(fit)$coef[2,4]
fit_label <- paste0("Linear Regression of SV breaks on CNV breaks:\n",
                   "Adjusted R = ", signif(corr, 3),
                     ", P = ", signif(pval, 3))
fit_label_grob <- grobTree(textGrob(fit_label, x=0.99, y=0.95, hjust=1))

# Create scatterplot
p <- merge %>%
  arrange(count_regions_any_conf) %>%
    # Sort by number of chromothripsis calls so that the samples with chromothripsis are more visible
  ggplot(aes(x = cnv_breaks_count, y = sv_breaks_count, fill = as.factor(count_regions_any_conf))) +
  geom_point(shape=21, size=2, color="black") + 
  scale_fill_manual(values=colors, name="# Chromothripsis\nRegions") +
  stat_smooth(aes(fill=NULL), method = "lm", col = "red", show.legend = FALSE) + 
  annotation_custom(fit_label_grob) +
  xlab("# CNV Breaks") + 
  ylab("# SV Breaks") 
p
ggsave(file.path(plots_dir, "count_chromothripsis_cnv_and_sv_breaks_scatterplot.png"), p)
# Also save a PDF version, to be used as a single panel in a multipanel figure
ggsave(file.path(plots_dir, "count_chromothripsis_cnv_and_sv_breaks_scatterplot.pdf"), p)
```

#### Stripchart: Chromothripsis regions vs. CNV breaks

```{r}
p <- merge %>%
  ggplot(aes(x = count_regions_any_conf_truncated, 
                               y = cnv_breaks_count, fill = count_regions_any_conf_truncated)) +
  geom_jitter(shape=21, size=2, color="black") +
  scale_fill_manual(values=colors) +
  theme(legend.position = "none") +
  geom_boxplot(fill=NA, outlier.shape=NA) +
  xlab("# Chromothripsis Regions") + 
  ylab("# CNV Breaks") 
p
ggsave(file.path(plots_dir, "count_chromothripsis_cnv_breaks_stripchart.png"), p)
```

#### Stripchart: Chromothripsis regions vs. SV breaks

```{r}
p <- merge %>%
  ggplot(aes(x = count_regions_any_conf_truncated, 
                               y = sv_breaks_count, fill = count_regions_any_conf_truncated)) +
  geom_jitter(shape=21, size=2, color="black") +
  scale_fill_manual(values=colors) +
  theme(legend.position = "none") +
  geom_boxplot(fill=NA, outlier.shape=NA) +
  xlab("# Chromothripsis Regions") + 
  ylab("# SV Breaks") 
p
ggsave(file.path(plots_dir, "count_chromothripsis_sv_breaks_stripchart.png"), p)
``` 

### Session Info

```{r}
sessionInfo()
```
